refactor(portal): extract playground YAML validation into testable module
Lift the hand-rolled YAML parser + tenant-config validator cluster
(parseYAML, parseDuration, validateDurationRange, isValidISO8601,
validateTenantConfig + KNOWN_METRIC_KEYS / RECEIVER_TYPES) out of
playground.jsx (777->418 LOC) into playground/validation.js
(PR-portal-16). This ~360 LOC of correctness-critical validation was
0%-covered; now 8 tests cover the parser, the duration contract, and the
validate flow (valid / no-tenants / bad-threshold / unknown-key warning /
routing-duration guardrail).

Kept playground-local on purpose: parseDuration/parseYAML here are NOT the
_common/validation/yaml-parser.js versions (parseDuration returns
{ms,value,unit} for s/m/h vs _common's seconds for s/m/h/d) — merging
would change behavior. The i18n `t` fallback moved with the cluster.

Behavior-preserving: logic moved verbatim, dist rebuilt, suite 204 -> 212.

Refs: TRK-230
Self-Review-Pass-2: 4-lens + refute; caught missing t-dep; _common dedup rejected; gates green
